got rid of configs b/c already done in hardware class. Also got rid of controls in sim classes b/c they already exist in the hardware class
