+++ src/main/java/frc/robot/subsystems/arm/ArmIO.java
@@ -61,70 +61,6 @@ public class ArmIO {
    */
   public void setArmVelocity(double velocity) {}
 
-  /**
-   * Gets the configuration used for the Talon FX motor controllers of the arm subsystem.
-   *
-   * <p>This method returns a Talon FX configuration with the following settings:
-   *
-   * <ul>
-   *   <li>Neutral mode: Brake
-   *   <li>Gravity type: Arm cosine
-   *   <li>Feedback device: Integrated sensor
-   *   <li>kP: {@link ArmConstants#ARM_MOTOR_kP}
-   *   <li>kI: {@link ArmConstants#ARM_MOTOR_kI}
-   *   <li>kD: {@link ArmConstants#ARM_MOTOR_kD}
-   *   <li>kG: {@link ArmConstants#ARM_MOTOR_kG}
-   *   <li>kV: {@link ArmConstants#ARM_MOTOR_kV}
-   *   <li>Current limit: 40A (CHANGE THIS VALUE OTHERWISE TORQUE MAY BE LIMITED/TOO HIGH)
-   * </ul>
-   *
-   * <p>These values may need to be changed based on the actual robot hardware and the desired
-   * behavior of the elevator.
-   *
-   * @return the configuration used for the Talon FX motor controllers of the arm subsystem
-   */
-  public SparkFlexConfig getArmConfig() {
-    SparkFlexConfig armConfig = new SparkFlexConfig();
-
-    armConfig
-        .idleMode(IdleMode.kBrake)
-        .inverted(true)
-        .smartCurrentLimit(ArmConstants.ARM_MOTOR_CURRENT_LIMIT)
-        .voltageCompensation(12.0);
-    armConfig
-        .absoluteEncoder
-        .zeroCentered(true)
-        .inverted(true)
-        .positionConversionFactor(1 / ArmSimConstants.GEARING)
-        .velocityConversionFactor(
-            2.0 / ArmSimConstants.GEARING); // TODO: figure out why we need this to be 2.0
-    armConfig
-        .closedLoop
-        .feedbackSensor(FeedbackSensor.kAbsoluteEncoder)
-        .positionWrappingEnabled(false)
-        .pid(
-            ArmConstants.ARM_MOTOR_POSITION_KP, ArmConstants.ARM_MOTOR_POSITION_KI,
-            ArmConstants.ARM_MOTOR_POSITION_KD, ClosedLoopSlot.kSlot0)
-        .pid(
-            ArmConstants.ARM_MOTOR_VELOCITY_KP, ArmConstants.ARM_MOTOR_VELOCITY_KI,
-            ArmConstants.ARM_MOTOR_VELOCITY_KD, ClosedLoopSlot.kSlot1);
-    armConfig
-        .signals
-        .absoluteEncoderPositionAlwaysOn(true)
-        .absoluteEncoderPositionPeriodMs((int) 10.0)
-        .absoluteEncoderVelocityAlwaysOn(true)
-        .absoluteEncoderVelocityPeriodMs(20)
-        .appliedOutputPeriodMs(20)
-        .busVoltagePeriodMs(20)
-        .outputCurrentPeriodMs(20);
-
-    // // added 3/6
-    // armConfig.softLimit.reverseSoftLimitEnabled(true);
-    // armConfig.softLimit.reverseSoftLimit(ArmConstants.SOFT_LIMIT);
-
-    return armConfig;
-  }
-
   public double getArmPosition(ArmIOInputs inputs) {
     return 0.0;
   }
