+++ src/main/java/frc/robot/subsystems/elevator/ElevatorIO.java
@@ -96,70 +96,5 @@ public class ElevatorIO {
     return -1;
   }
 
-  /**
-   * Gets the configuration used for the Talon FX motor controllers of the elevator subsystem.
-   *
-   * <p>This method returns a Talon FX configuration with the following settings:
-   *
-   * <ul>
-   *   <li>Neutral mode: Brake
-   *   <li>Gravity type: Elevator static
-   *   <li>Feedback device: Integrated sensor
-   *   <li>kP: {@link ElevatorConstants#ELEVATOR_MOTOR_kP}
-   *   <li>kI: {@link ElevatorConstants#ELEVATOR_MOTOR_kI}
-   *   <li>kD: {@link ElevatorConstants#ELEVATOR_MOTOR_kD}
-   *   <li>kG: {@link ElevatorConstants#ELEVATOR_MOTOR_kG}
-   *   <li>kV: {@link ElevatorConstants#ELEVATOR_MOTOR_kV}
-   *   <li>Current limit: 40A (CHANGE THIS VALUE OTHERWISE TORQUE MAY BE LIMITED/TOO HIGH)
-   * </ul>
-   *
-   * <p>These values may need to be changed based on the actual robot hardware and the desired
-   * behavior of the elevator.
-   *
-   * @return the configuration used for the Talon FX motor controllers of the elevator subsystem
-   */
-  public TalonFXConfiguration getConfiguration(int motorNum) {
-    var config = new TalonFXConfiguration();
-    config.Voltage.PeakForwardVoltage = 16;
-    config.MotorOutput.NeutralMode = NeutralModeValue.Brake;
-
-    // Slot 0 is position
-    config.Slot0.GravityType = GravityTypeValue.Elevator_Static;
-    config.Slot0.kP = ElevatorConstants.ElevatorPositionConstants.ELEVATOR_MOTOR_kP;
-    config.Slot0.kI = ElevatorConstants.ElevatorPositionConstants.ELEVATOR_MOTOR_kI;
-    config.Slot0.kD = ElevatorConstants.ElevatorPositionConstants.ELEVATOR_MOTOR_kD;
-    config.Slot0.kG = ElevatorConstants.ElevatorPositionConstants.ELEVATOR_MOTOR_kG;
-    config.Slot0.kV = ElevatorConstants.ElevatorPositionConstants.ELEVATOR_MOTOR_kV;
-
-    // Slot 1 is velocity
-
-    config.Slot1.GravityType = GravityTypeValue.Elevator_Static;
-    config.Slot1.kP = ElevatorConstants.ElevatorVelocityConstants.ELEVATOR_MOTOR_kP;
-    config.Slot1.kI = ElevatorConstants.ElevatorVelocityConstants.ELEVATOR_MOTOR_kI;
-    config.Slot1.kD = ElevatorConstants.ElevatorVelocityConstants.ELEVATOR_MOTOR_kD;
-    config.Slot1.kG = ElevatorConstants.ElevatorVelocityConstants.ELEVATOR_MOTOR_kG;
-    config.Slot1.kV = ElevatorConstants.ElevatorVelocityConstants.ELEVATOR_MOTOR_kV;
-
-    config.Slot2.GravityType = GravityTypeValue.Elevator_Static;
-    config.Slot2.kP = ElevatorConstants.ElevatorStowPresetConstants.ELEVATOR_MOTOR_kP;
-    config.Slot2.kI = ElevatorConstants.ElevatorStowPresetConstants.ELEVATOR_MOTOR_kI;
-    config.Slot2.kD = ElevatorConstants.ElevatorStowPresetConstants.ELEVATOR_MOTOR_kD;
-    config.Slot2.kG = ElevatorConstants.ElevatorStowPresetConstants.ELEVATOR_MOTOR_kG;
-    config.Slot2.kV = ElevatorConstants.ElevatorStowPresetConstants.ELEVATOR_MOTOR_kV;
-
-    config.SoftwareLimitSwitch.ForwardSoftLimitEnable = true;
-    config.SoftwareLimitSwitch.ForwardSoftLimitThreshold = 22.7;
-    config.SoftwareLimitSwitch.ReverseSoftLimitEnable = true;
-    config.SoftwareLimitSwitch.ReverseSoftLimitThreshold = 0.1;
-
-    var currentConfig = new CurrentLimitsConfigs();
-    currentConfig.StatorCurrentLimitEnable = true;
-    // CHANGE THIS VALUE OTHERWISE TORQUE MAY BE LIMITED/TOO HIGH
-    currentConfig.StatorCurrentLimit = 140;
-    config.CurrentLimits = currentConfig;
-    if (motorNum == 2) config.MotorOutput.Inverted = InvertedValue.Clockwise_Positive;
-    return config;
-  }
-
   public void elevatorOpenLoop(Voltage voltage) {}
 }
