+++ src/main/java/frc/robot/subsystems/elevator/ElevatorIOTalonFX.java
@@ -4,8 +4,14 @@ import static frc.robot.util.PhoenixUtil.tryUntilOk;
 
 import com.ctre.phoenix6.BaseStatusSignal;
 import com.ctre.phoenix6.StatusSignal;
+import com.ctre.phoenix6.configs.CurrentLimitsConfigs;
+import com.ctre.phoenix6.configs.TalonFXConfiguration;
 import com.ctre.phoenix6.controls.StrictFollower;
 import com.ctre.phoenix6.hardware.ParentDevice;
+import com.ctre.phoenix6.signals.GravityTypeValue;
+import com.ctre.phoenix6.signals.InvertedValue;
+import com.ctre.phoenix6.signals.NeutralModeValue;
+
 import edu.wpi.first.math.filter.Debouncer;
 import edu.wpi.first.math.util.Units;
 import edu.wpi.first.units.measure.Angle;
@@ -69,6 +75,71 @@ public class ElevatorIOTalonFX extends ElevatorIO {
                 elevator2Current));
     ParentDevice.optimizeBusUtilizationForAll(elevatorMotor1, elevatorMotor2);
     elevatorMotor2.setControl(new StrictFollower(elevatorMotor1.getDeviceID()));
+  }
+
+  /**
+   * Gets the configuration used for the Talon FX motor controllers of the elevator subsystem.
+   *
+   * <p>This method returns a Talon FX configuration with the following settings:
+   *
+   * <ul>
+   *   <li>Neutral mode: Brake
+   *   <li>Gravity type: Elevator static
+   *   <li>Feedback device: Integrated sensor
+   *   <li>kP: {@link ElevatorConstants#ELEVATOR_MOTOR_kP}
+   *   <li>kI: {@link ElevatorConstants#ELEVATOR_MOTOR_kI}
+   *   <li>kD: {@link ElevatorConstants#ELEVATOR_MOTOR_kD}
+   *   <li>kG: {@link ElevatorConstants#ELEVATOR_MOTOR_kG}
+   *   <li>kV: {@link ElevatorConstants#ELEVATOR_MOTOR_kV}
+   *   <li>Current limit: 40A (CHANGE THIS VALUE OTHERWISE TORQUE MAY BE LIMITED/TOO HIGH)
+   * </ul>
+   *
+   * <p>These values may need to be changed based on the actual robot hardware and the desired
+   * behavior of the elevator.
+   *
+   * @return the configuration used for the Talon FX motor controllers of the elevator subsystem
+   */
+  public TalonFXConfiguration getConfiguration(int motorNum) {
+    var config = new TalonFXConfiguration();
+    config.Voltage.PeakForwardVoltage = 16;
+    config.MotorOutput.NeutralMode = NeutralModeValue.Brake;
+
+    // Slot 0 is position
+    config.Slot0.GravityType = GravityTypeValue.Elevator_Static;
+    config.Slot0.kP = ElevatorConstants.ElevatorPositionConstants.ELEVATOR_MOTOR_kP;
+    config.Slot0.kI = ElevatorConstants.ElevatorPositionConstants.ELEVATOR_MOTOR_kI;
+    config.Slot0.kD = ElevatorConstants.ElevatorPositionConstants.ELEVATOR_MOTOR_kD;
+    config.Slot0.kG = ElevatorConstants.ElevatorPositionConstants.ELEVATOR_MOTOR_kG;
+    config.Slot0.kV = ElevatorConstants.ElevatorPositionConstants.ELEVATOR_MOTOR_kV;
+
+    // Slot 1 is velocity
+
+    config.Slot1.GravityType = GravityTypeValue.Elevator_Static;
+    config.Slot1.kP = ElevatorConstants.ElevatorVelocityConstants.ELEVATOR_MOTOR_kP;
+    config.Slot1.kI = ElevatorConstants.ElevatorVelocityConstants.ELEVATOR_MOTOR_kI;
+    config.Slot1.kD = ElevatorConstants.ElevatorVelocityConstants.ELEVATOR_MOTOR_kD;
+    config.Slot1.kG = ElevatorConstants.ElevatorVelocityConstants.ELEVATOR_MOTOR_kG;
+    config.Slot1.kV = ElevatorConstants.ElevatorVelocityConstants.ELEVATOR_MOTOR_kV;
+
+    config.Slot2.GravityType = GravityTypeValue.Elevator_Static;
+    config.Slot2.kP = ElevatorConstants.ElevatorStowPresetConstants.ELEVATOR_MOTOR_kP;
+    config.Slot2.kI = ElevatorConstants.ElevatorStowPresetConstants.ELEVATOR_MOTOR_kI;
+    config.Slot2.kD = ElevatorConstants.ElevatorStowPresetConstants.ELEVATOR_MOTOR_kD;
+    config.Slot2.kG = ElevatorConstants.ElevatorStowPresetConstants.ELEVATOR_MOTOR_kG;
+    config.Slot2.kV = ElevatorConstants.ElevatorStowPresetConstants.ELEVATOR_MOTOR_kV;
+
+    config.SoftwareLimitSwitch.ForwardSoftLimitEnable = true;
+    config.SoftwareLimitSwitch.ForwardSoftLimitThreshold = 22.7;
+    config.SoftwareLimitSwitch.ReverseSoftLimitEnable = true;
+    config.SoftwareLimitSwitch.ReverseSoftLimitThreshold = 0.1;
+
+    var currentConfig = new CurrentLimitsConfigs();
+    currentConfig.StatorCurrentLimitEnable = true;
+    // CHANGE THIS VALUE OTHERWISE TORQUE MAY BE LIMITED/TOO HIGH
+    currentConfig.StatorCurrentLimit = 140;
+    config.CurrentLimits = currentConfig;
+    if (motorNum == 2) config.MotorOutput.Inverted = InvertedValue.Clockwise_Positive;
+    return config;
   }
 
   @Override
