+++ src/main/java/frc/robot/subsystems/endeffector/EndEffectorIOSim.java
@@ -3,7 +3,6 @@ package frc.robot.subsystems.endeffector;
 import com.ctre.phoenix6.sim.TalonFXSimState;
 import edu.wpi.first.math.system.plant.DCMotor;
 import edu.wpi.first.math.system.plant.LinearSystemId;
-import edu.wpi.first.units.measure.Voltage;
 import edu.wpi.first.wpilibj.RobotController;
 import edu.wpi.first.wpilibj.simulation.FlywheelSim;
 import frc.robot.subsystems.SimMechanism;
@@ -21,7 +20,6 @@ public class EndEffectorIOSim extends EndEffectorIOTalonFX implements SimMechani
 
   public EndEffectorIOSim() {
     super();
-    endEffectorMotor.getConfigurator().apply(getEndEffectorConfiguration());
     initSimVoltage();
   }
 
@@ -43,19 +41,19 @@ public class EndEffectorIOSim extends EndEffectorIOTalonFX implements SimMechani
         physicsSim.getAngularVelocityRadPerSec() * EndEffectorSimConstants.GEARING);
   }
 
-  @Override
-  public void setEndEffectorVelocity(double velocity) {
-    endEffectorMotor.setControl(
-        endEffectorVelocityRequest.withVelocity(velocity * EndEffectorSimConstants.GEARING));
-  }
+  // @Override
+  // public void setEndEffectorVelocity(double velocity) {
+  //   endEffectorMotor.setControl(
+  //       endEffectorVelocityRequest.withVelocity(velocity * EndEffectorSimConstants.GEARING));
+  // }
 
   @Override
   public double[] getCurrents() {
     return new double[] {physicsSim.getCurrentDrawAmps()};
   }
 
-  @Override
-  public void endEffectorOpenLoop(Voltage voltage) {
-    endEffectorMotor.setVoltage(voltage.magnitude());
-  }
+  // @Override
+  // public void endEffectorOpenLoop(Voltage voltage) {
+  //   endEffectorMotor.setVoltage(voltage.magnitude());
+  // }
 }
